Fix bugs of CL and add matmul of CL(ROW) and CL(Row+Lmem)

+++ matmul_neon_cl/src/naive_cl_impl.cl
@@ -1,15 +1,15 @@
-__kernel void matmul(__global int* a,
-                            __global int* b,
-                            __global int* c,
-                            const int M, 
-                            const int N, 
-                            const int K){
+__kernel void matmul(__global float* a,
+                        __global float* b,
+                        __global float* c,
+                        const int M, 
+                        const int N, 
+                        const int K){
     
     int i = get_global_id(0);
     int j = get_global_id(1);
     int index = (i * N) + j;
 
-    int sum = 0;
+    float sum = 0;
     for(int k = 0; k < K; k++){
         sum += a[i + k * M] * b[k * N + j];
     }
